requirements.txt
annotated-types==0.6.0
anyio==4.3.0
av==11.0.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
ctranslate2==4.1.0
distro==1.9.0
exceptiongroup==1.2.0
faster-whisper==1.0.1
filelock==3.13.3
Flask==3.0.2
Flask-Cors==4.0.0
flatbuffers==24.3.25
fsspec==2024.3.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mecab-python3==1.0.8
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
onnxruntime==1.17.1
openai==1.16.2
packaging==24.0
protobuf==5.26.1
psycopg2-binary==2.9.9
pyasn1==0.6.0
pydantic==2.6.4
pydantic_core==2.16.3
pysmb==1.2.9.1
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.1
sympy==1.12
tokenizers==0.15.2
torch==2.2.2
tqdm==4.66.2
typing_extensions==4.10.0
unidic-lite==1.0.8
urllib3==2.2.1
Werkzeug==3.0.2